requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
Pillow
numpy
tortoise-orm
asyncpg